requirements.txt
apache-airflow
nltk
cached_property